requirements.txt
faker
snowflake-connector-python
snowflake-snowpark-python
streamlit
